Command Pattern
exchange web service (EWS) is the system that sends emails to list of users. The command pattern will have an instance of the receiver and it will need all the elements that are going to be used in the implementation of the commandObject's execute() method. In there we will have constructor that are needed by the executor.
In our main class we need to have an instance of the commandClass